Add test for listener invocation on updating imported data
